Made a temporary view and it works.
Model, View, and Control were all attached where needed in TestDriver2 and it works. It is able to read and write data, add new data, and display the data. Now we need to work on the removal of data and making View the way it should be. I'm also trying to set gitignore to ignore csv files.